indui 1.3.0
This fixes up misleading errors from styled-jsx which were
actually caused by using identifiers that are HTML elements.
Moving on, the relase splits up the TextLine component into
an individual Input and the former TextLine component. They
are closely related in terms of appearance, and refactoring
will probably happen - or they may be merged again if we can
find a sensible API.